Add files via upload
 Sunglasses and Goggles Showcase Website:
I made a website of Sunglasses and Goggles using HTML and CSS as well as using Flask. 
"A responsive sunglasses and goggles showcase website built with Flask, HTML, and CSS, featuring a simulated login system, a dynamic gallery, and a modern, user-friendly design. This project demonstrates core web development concepts and is tailored for educational purposes."